refactor: extract bridge readiness logic into a separate helper module
