+++ apps/wso2-digital-wallet/src/components/Home/RecentActivities.jsx
@@ -27,6 +27,7 @@ import {
 } from '../../constants/strings';
 import { getLocalDataAsync } from '../../helpers/storage';
 import { getRecentTransactions } from '../../services/blockchain.service';
+import { waitForBridge } from '../../helpers/bridge';
 
 function RecentActivities() {
   const [walletAddress, setWalletAddress] = useState("");
@@ -36,20 +37,6 @@ function RecentActivities() {
   const [isFetchingInBackground, setIsFetchingInBackground] = useState(false);
 
   const orangeColor = "#ff7300";
-
-  const checkBridgeReady = () => {
-    return window.nativebridge && window.ReactNativeWebView;
-  };
-
-  const waitForBridge = async (maxWaitTime = 5000) => {
-    const startTime = Date.now();
-    
-    while (!checkBridgeReady() && (Date.now() - startTime) < maxWaitTime) {
-      await new Promise(resolve => setTimeout(resolve, 100));
-    }
-    
-    return checkBridgeReady();
-  };
 
   const fetchWalletAddress = async () => {
     try {
