+++ apps/wso2-digital-wallet/src/pages/Home/Home.js
@@ -40,6 +40,7 @@ import { getLocalDataAsync } from '../../helpers/storage';
 import {
   getWalletBalanceByWalletAddress,
 } from '../../services/blockchain.service';
+import { waitForBridge } from '../../helpers/bridge';
 
 function Home() {
   const navigate = useNavigate();
@@ -114,20 +115,6 @@ function Home() {
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [walletAddress]);
-
-  const checkBridgeReady = () => {
-    return window.nativebridge && window.ReactNativeWebView;
-  };
-
-  const waitForBridge = async (maxWaitTime = 5000) => {
-    const startTime = Date.now();
-    
-    while (!checkBridgeReady() && (Date.now() - startTime) < maxWaitTime) {
-      await new Promise(resolve => setTimeout(resolve, 100));
-    }
-    
-    return checkBridgeReady();
-  };
 
   const fetchCurrentTokenBalance = async () => {
     try {
