+++ apps/wso2-digital-wallet/src/pages/SendAssets/SendAssets.js
@@ -23,6 +23,7 @@ import {
 import { getLocalDataAsync, saveLocalDataAsync } from "../../helpers/storage";
 import { STORAGE_KEYS, DEFAULT_WALLET_ADDRESS } from "../../constants/configs";
 import { getWalletBalanceByWalletAddress } from "../../services/blockchain.service";
+import { waitForBridge } from "../../helpers/bridge";
 
 function SendAssets() {
   const navigate = useNavigate();
@@ -53,20 +54,6 @@ function SendAssets() {
       console.log(`${ERROR_RETRIEVE_WALLET_ADDRESS} - ${error}`);
       messageApi.error(ERROR_RETRIEVE_WALLET_ADDRESS);
     }
-  };
-
-  const checkBridgeReady = () => {
-    return window.nativebridge && window.ReactNativeWebView;
-  };
-
-  const waitForBridge = async (maxWaitTime = 5000) => {
-    const startTime = Date.now();
-    
-    while (!checkBridgeReady() && (Date.now() - startTime) < maxWaitTime) {
-      await new Promise(resolve => setTimeout(resolve, 100));
-    }
-    
-    return checkBridgeReady();
   };
 
   const fetchCurrentTokenBalance = async () => {
